backend-vton/requirements.txt
fastapi
uvicorn
python-multipart
Pillow

# Add these when you are ready to implement the real OOTDiffusion Model and have a GPU:
# torch
# torchvision
# diffusers
# transformers
# accelerate
